migrated byte formatting in strings to util library
