+++ src/lib/scrapers/strategies/base.ts
@@ -1,6 +1,7 @@
 import { ChildProcess, spawn } from 'child_process';
 import { ScraperOptions, ScrapeResult } from '@/lib/scrapers/types';
 import { ScrapeLimits } from '@/lib/scrapers/runner';
+import { parseSizeToBytes, formatBytes } from '@/lib/utils/format';
 
 export abstract class BaseScraperStrategy {
     protected basePath: string;
@@ -79,7 +80,7 @@ export abstract class BaseScraperStrategy {
             console.log(`[ScraperStrategy] Reached post limit of ${this.limits.stopAfterPosts}. Waiting for finalization...`);
             this.intentionalStop = true;
             this.stopInitiated = true;
-            
+
             // Give the scraper a moment to finish the current post's download/file operations
             // before we kill it. This prevents the 0-byte .part file issue.
             setTimeout(() => {
@@ -96,35 +97,8 @@ export abstract class BaseScraperStrategy {
         }
     }
 
-    protected parseSizeStringToBytes(sizeStr: string): number {
-        if (!sizeStr) return 0;
-        const match = sizeStr.trim().match(/^([\d.]+)\s*([A-Za-z]*)$/);
-        if (!match) return 0;
-
-        const value = parseFloat(match[1]);
-        const unit = match[2].toLowerCase();
-
-        if (!unit) return Math.floor(value);
-
-        const multipliers: { [key: string]: number } = {
-            'b': 1, 'k': 1024, 'kb': 1024, 'kib': 1024, 'm': 1024 * 1024,
-            'mb': 1024 * 1024, 'mib': 1024 * 1024, 'g': 1024 * 1024 * 1024,
-            'gb': 1024 * 1024 * 1024, 'gib': 1024 * 1024 * 1024,
-            't': 1024 * 1024 * 1024 * 1024, 'tb': 1024 * 1024 * 1024 * 1024,
-            'tib': 1024 * 1024 * 1024 * 1024,
-        };
-
-        return Math.floor(value * (multipliers[unit] || 0));
-    }
-
-    protected formatBytes(bytes: number): string {
-        if (bytes === 0) return '0B';
-        if (bytes < 1024) return bytes + 'B';
-        const k = 1024;
-        const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
-        const i = Math.floor(Math.log(bytes) / Math.log(k));
-        return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + sizes[i];
-    }
+    parseSizeToBytes = parseSizeToBytes;
+    formatBytes = formatBytes;
 
     public getFinalResult(success: boolean, errorMsg?: string): ScrapeResult {
         if (this.options.onProgress) {
